% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{run_neighbournet_networkx}
\alias{run_neighbournet_networkx}
\title{Run NeighbourNet (alias) and return a networkx-like list.}
\usage{
run_neighbournet_networkx(
  x,
  flip_y = TRUE,
  labels = NULL,
  max_iterations = 5000,
  ordering_method = NULL
)
}
\description{
Run NeighbourNet (alias) and return a networkx-like list.
}
