% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{set_fastnnt_threads}
\alias{set_fastnnt_threads}
\title{Set the global thread count for fastnnt (call once per session).}
\usage{
set_fastnnt_threads(threads)
}
\description{
Set the global thread count for fastnnt (call once per session).
}
